+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="1.0.3"></a>
+## 1.0.3 (2021-01-30)
+
+
+### Features
+* **core:** add `organizations` and `projects` query into the `/stream` endpoint ([CS-403](https://jira.rfcx.org/browse/CS-403))
+
+
 <a name="1.0.1"></a>
 ## 1.0.1 (2021-01-30)
 
@@ -5,7 +13,6 @@
 ### Features
 * **core:** update events endpoints to support `systemUser` ([CS-105](https://jira.rfcx.org/browse/CS-105))
 * **core:** update classifier `last_executed_at` when saving detections ([CS-368](https://jira.rfcx.org/browse/CS-368))
-* **core:** add `organizations` and `projects` query into the `/stream` endpoint ([CS-403](https://jira.rfcx.org/browse/CS-403))
 
 
 <a name="1.0.0"></a>